ai/requirements.txt
# Google Generative AI SDK
google-generativeai==0.3.2

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Image processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
